+++ stack-core/src/main/java/com/inductiveautomation/opcua/stack/core/types/structured/RedundantServerDataType.java
@@ -6,6 +6,7 @@ import com.inductiveautomation.opcua.stack.core.serialization.UaDecoder;
 import com.inductiveautomation.opcua.stack.core.serialization.UaEncoder;
 import com.inductiveautomation.opcua.stack.core.serialization.UaStructure;
 import com.inductiveautomation.opcua.stack.core.types.builtin.NodeId;
+import com.inductiveautomation.opcua.stack.core.types.builtin.unsigned.UByte;
 import com.inductiveautomation.opcua.stack.core.types.enumerated.ServerState;
 
 public class RedundantServerDataType implements UaStructure {
@@ -15,29 +16,41 @@ public class RedundantServerDataType implements UaStructure {
     public static final NodeId XmlEncodingId = Identifiers.RedundantServerDataType_Encoding_DefaultXml;
 
     protected final String _serverId;
-    protected final Short _serviceLevel;
+    protected final UByte _serviceLevel;
     protected final ServerState _serverState;
 
-    public RedundantServerDataType(String _serverId, Short _serviceLevel, ServerState _serverState) {
+    public RedundantServerDataType(String _serverId, UByte _serviceLevel, ServerState _serverState) {
         this._serverId = _serverId;
         this._serviceLevel = _serviceLevel;
         this._serverState = _serverState;
     }
 
-    public String getServerId() { return _serverId; }
+    public String getServerId() {
+        return _serverId;
+    }
 
-    public Short getServiceLevel() { return _serviceLevel; }
+    public UByte getServiceLevel() {
+        return _serviceLevel;
+    }
 
-    public ServerState getServerState() { return _serverState; }
+    public ServerState getServerState() {
+        return _serverState;
+    }
 
     @Override
-    public NodeId getTypeId() { return TypeId; }
+    public NodeId getTypeId() {
+        return TypeId;
+    }
 
     @Override
-    public NodeId getBinaryEncodingId() { return BinaryEncodingId; }
+    public NodeId getBinaryEncodingId() {
+        return BinaryEncodingId;
+    }
 
     @Override
-    public NodeId getXmlEncodingId() { return XmlEncodingId; }
+    public NodeId getXmlEncodingId() {
+        return XmlEncodingId;
+    }
 
 
     public static void encode(RedundantServerDataType redundantServerDataType, UaEncoder encoder) {
@@ -48,7 +61,7 @@ public class RedundantServerDataType implements UaStructure {
 
     public static RedundantServerDataType decode(UaDecoder decoder) {
         String _serverId = decoder.decodeString("ServerId");
-        Short _serviceLevel = decoder.decodeByte("ServiceLevel");
+        UByte _serviceLevel = decoder.decodeByte("ServiceLevel");
         ServerState _serverState = decoder.decodeSerializable("ServerState", ServerState.class);
 
         return new RedundantServerDataType(_serverId, _serviceLevel, _serverState);
